requirements.txt
numpy==1.23.0
pillow==9.5.0
opencv-python==4.8.0.74